fix: max file name size constant overflow

+++ src/include/Constants.h
@@ -8,7 +8,7 @@
 #include <cstdint>
 
 constexpr uint8_t kDefaultBacklog = 10;
-constexpr uint8_t kMaxFileNameSize = 4096;
+constexpr uint16_t kMaxFileNameSize = 4096;
 constexpr uint64_t kMaxFileSize = 10e12;
 
 #endif //CONSTANTS_H
